Add watchdog to get rid of interfering signals
Switching on the light in the room would sometimes
cause LEDs to light up permanently.
The watchdog switches all LEDs back out if no
alarm is rining. This will temporary break
lullabies, as they are currently not registered as
running alarms and the watchdog will therefore
be oblivious to them.